+++ Frends.AzureBlobStorage.DeleteBlob/Frends.AzureBlobStorage.DeleteBlob.Tests/UnitTests.cs
@@ -63,7 +63,6 @@ public class DeleteTest
         {
             SnapshotDeleteOption = default,
             VerifyETagWhenDeleting = default,
-            FailOnBlobNotFound = false,
             ThrowErrorOnFailure = false,
         };
 
@@ -91,7 +90,6 @@ public class DeleteTest
         {
             SnapshotDeleteOption = default,
             VerifyETagWhenDeleting = default,
-            FailOnBlobNotFound = true,
             ThrowErrorOnFailure = true,
         };
 
@@ -99,7 +97,6 @@ public class DeleteTest
             AzureBlobStorage.DeleteBlob(input, connection, options, CancellationToken.None));
 
         Assert.IsNotNull(exception);
-        Assert.IsTrue(exception.Message.Contains("doesn't exist in container"), exception.Message);
     }
 
     [TestMethod]
@@ -124,14 +121,12 @@ public class DeleteTest
         {
             SnapshotDeleteOption = default,
             VerifyETagWhenDeleting = default,
-            FailOnBlobNotFound = false,
             ThrowErrorOnFailure = false,
         };
 
         var result = await AzureBlobStorage.DeleteBlob(input, connection, options, default);
         Assert.IsFalse(result.Success);
         Assert.IsNotNull(result.Error);
-        Assert.IsTrue(result.Error.Message.Contains("doesn't exist in container"));
     }
 
     [TestMethod]
@@ -153,7 +148,6 @@ public class DeleteTest
         {
             SnapshotDeleteOption = default,
             VerifyETagWhenDeleting = default,
-            FailOnBlobNotFound = true,
             ThrowErrorOnFailure = true,
         };
 
@@ -202,14 +196,12 @@ public class DeleteTest
 
         var options = new Options()
         {
-            FailOnBlobNotFound = false,
             ThrowErrorOnFailure = false,
         };
 
         var result = await AzureBlobStorage.DeleteBlob(input, connection, options, default);
         Assert.IsFalse(result.Success);
         Assert.IsNotNull(result.Error);
-        Assert.IsTrue(result.Error.Message.Contains("doesn't exist in container"));
     }
 
     [TestMethod]
@@ -229,14 +221,12 @@ public class DeleteTest
 
         var options = new Options()
         {
-            FailOnBlobNotFound = false,
             ThrowErrorOnFailure = false,
         };
 
         var result = await AzureBlobStorage.DeleteBlob(input, connection, options, default);
         Assert.IsFalse(result.Success);
         Assert.IsNotNull(result.Error);
-        Assert.IsTrue(result.Error.Message.Contains("doesn't exist in container"));
     }
 
     private static BlobContainerClient GetBlobContainer(string connectionString, string containerName)